Revert "Move stuff changes"
