--- app/api/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-bp = Blueprint('api', __name__)
-
-from app.api import routes
