+++ configs.py
@@ -1,24 +1,5 @@
-
-import sys, os
-
-def get_sys_exec_root_or_drive():
-    path = sys.executable
-    while os.path.split(path)[1]:
-        path = os.path.split(path)[0]
-    return path
-
-
-if os.environ.get('DEVELOPMENT'):
-    alternative_db = 'sqlite:///' + os.path.join(get_sys_exec_root_or_drive(), 'data.sqlite')
-
-    print(alternative_db    )
-
-else:
-    if not os.environ.get('SQLALCHEMY_DATABASE_URI'):
-        raise KeyError("Application requires 'SQLALCHEMY_DATABASE_URL' to run in non DEVELOPMENT")
-
+import os
 
 class Config:
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-    SQLALCHEMY_DATABASE_URI = os.environ.get('SQLALCHEMY_DATABASE_URI') or alternative_db
-    SQL_LITE_DEFAULT = False or (alternative_db != None)
+    SQLALCHEMY_DATABASE_URI = os.environ.get('SQLALCHEMY_DATABASE_URI')
+    SQLALCHEMY_TRACK_MODIFICATIONS = True
